requirements.txt
numpy
scipy
easydict
h5py
https://download.pytorch.org/whl/cu90/torch-0.4.0-cp35-cp35m-linux_x86_64.whl
torchvision
matplotlib
vtk